requirements.txt
gradio>=4.0
tensorflow
pillow
numpy
